Rewrite synchronizeData aka packetReframe
The original concept was to only use the CRC to identify frames.  The
issue here is that we want to use FEC so when dropping all broken/crc
fail packets FEC is not aplicable.

So we use the stats for 100 consecutive packets with CRC fail to begin
searching for a CRC valid packet in the buffer.

In case the error rate increases and we have 100 consecutive packets
with CRC fail our FEC will fail anyway so dropping bytes/packets and
resetting state will do no harm and resynchronize on the first CRC valid
packet.

In case we lost sync with the packet stream because of chip reset, MSC
Data Buffer Over/Underrun we will resynchronize after 100 packets. Worst
case thats 96Byte packets with a 8KBit/s Packet Data Stream which takes
about ~10 Seconds to try to find packet only by CRC.  Best case at
8KBit/s its 2.4 seconds.

At 24 byte packet the FEC Frame (2256 bytes) is 94 packets.